+++ src/app/app.component.ts
@@ -1,7 +1,7 @@
 import {ChangeDetectionStrategy, Component} from '@angular/core';
 import {RouterOutlet} from '@angular/router';
 import {ProductComponent} from "./product/product.component";
-import {CardListComponent} from "./card/card.component";
+import {CardListComponent} from "./card/cardList.component";
 import {NgForOf} from "@angular/common";
 import {products} from "./constants";
 import {Card, Product} from "./interfaces";
@@ -20,30 +20,28 @@ export class AppComponent {
     protected _cardList: Card[] = [];
 
     protected onProductSelected(product: Product): void {
-        const card = this._cardList.find(element => element.id === product.id);
-        if (!card) {
+        if (!this._cardList.some(element => element.id === product.id)) {
             this.addProductCard(product);
-            this.incrementProductCard(product);
-        } else {
-            this.incrementProductCard(product);
         }
+        this.incrementProductCard(product);
     }
 
     protected productTrackBy(index: number, product: { id: number }): number {
         return product.id;
     }
 
-    private addProductCard(product: Product): void {
+    private addProductCard(product: Product): Card {
         this._cardList = [...this._cardList, {
             'id': product.id,
             'quantity': 0
         }];
+        return this._cardList.at(-1);
     }
 
     private incrementProductCard(product: Product): void {
-        this._cardList =this._cardList.map(element => {
+        this._cardList = this._cardList.map(element => {
             if (element.id === product.id) {
-                return {...element, quantity: element.quantity+1};
+                return {...element, quantity: element.quantity + 1};
             }
             return element;
         })
