+++ src/app/card/cardList.component.css
