+++ src/app/product/product.component.html
@@ -2,5 +2,5 @@
   <img [src]="image" alt="{{ name }}" class="product-image">
   <h3 class="product-name"> {{ name }} </h3>
   <p class="product-price"> {{ price }}€ </p>
-  <button class="add-to-card-button" (click)="onProductClick()">Ajouter au panier</button>
+  <button class="add-to-card-button" (click)="_onProductClick()">Ajouter au panier</button>
 </div>
