+++ src/app/product/product.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
+import {Component, EventEmitter, Input, Output} from '@angular/core';
 
 @Component({
   selector: 'app-product',
@@ -7,13 +7,19 @@ import { Component, Input, Output, EventEmitter } from '@angular/core';
   standalone: true,
 })
 export class ProductComponent {
-  @Input() public name!: string;
-  @Input() public image!: string;
-  @Input() public price!: number;
+  @Input()
+  public readonly name: string;
 
-  @Output() public productSelected = new EventEmitter<void>();
+  @Input()
+  public readonly image: string;
 
-  protected onProductClick(): void {
+  @Input()
+  public readonly  price: number;
+
+  @Output()
+  public readonly productSelected = new EventEmitter<void>();
+
+  protected _onProductClick(): void {
     this.productSelected.emit();
   }
 }
